[#5] refactor : common패키지의 Errorcode 통합
