+++ src/main/java/com/localcoupon/couponservice/common/enums/CommonErrorCode.java
@@ -1,16 +1,16 @@
-package com.localcoupon.couponservice.common;
+package com.localcoupon.couponservice.common.enums;
 
 import com.localcoupon.couponservice.common.exception.ErrorCode;
 import org.springframework.http.HttpStatus;
 
 public enum CommonErrorCode implements ErrorCode {
-    BAD_REQUEST("잘못된 요청 입니다.", HttpStatus.BAD_REQUEST),
+    JSON_SERIALIZE_ERROR("직렬화에 실패했습니다.", HttpStatus.BAD_REQUEST),
+    GEO_LOCATION_ERROR("위치 정보가 올바르지 않습니다.", HttpStatus.BAD_REQUEST),
+    DUPLICATE_ERROR("중복 요청은 불가능합니다.", HttpStatus.BAD_REQUEST),
+    BAD_REQUEST("중복 요청은 불가능합니다.", HttpStatus.BAD_REQUEST),
+    ENTITY_NOT_FOUND_ERROR("데이터가 없습니다.", HttpStatus.NO_CONTENT),
     SERVER_ERROR("서버 에러 입니다.", HttpStatus.INTERNAL_SERVER_ERROR),
-    DATABASE_ERROR("데이터베이스 처리 중 오류가 발생했습니다.", HttpStatus.INTERNAL_SERVER_ERROR),
-    JSON_SERIALIZE_ERROR("JSON 직렬화에 실패했습니다.", HttpStatus.INTERNAL_SERVER_ERROR),
-    GEO_LOCATION_ERROR("좌표정보가 올바르지 않습니다.", HttpStatus.BAD_REQUEST),
-    ENTITY_NOT_FOUND_ERROR("엔티티 정보가 없습니다.", HttpStatus.NO_CONTENT);
-
+    DATABASE_ERROR("데이터베이스 처리 중 오류가 발생했습니다.", HttpStatus.INTERNAL_SERVER_ERROR);
     private final String message;
     private final HttpStatus status;
