+++ src/main/java/com/localcoupon/couponservice/common/exception/CommonException.java
@@ -1,0 +1,9 @@
+package com.localcoupon.couponservice.common.exception;
+
+import com.localcoupon.couponservice.common.enums.CommonErrorCode;
+
+public class CommonException extends BaseException {
+    public CommonException(CommonErrorCode errorCode) {
+        super(errorCode);
+    }
+}
