--- src/main/java/com/localcoupon/couponservice/common/exception/GeoException.java
@@ -1,7 +1,0 @@
-package com.localcoupon.couponservice.common.exception;
-
-public class GeoException extends BaseException{
-    public GeoException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
