--- src/main/java/com/localcoupon/couponservice/common/exception/JsonSerializeException.java
@@ -1,8 +1,0 @@
-package com.localcoupon.couponservice.common.exception;
-
-public class JsonSerializeException extends BaseException {
-    public JsonSerializeException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
-
